StickerView 인터페이스와 해당 인터페이스 구현체 코드 분리
